backend/requirements.txt
flask
flask-cors
pyyaml
azure-ai-inference